requirements.txt
openai
epicbox
ipykernel
matplotlib
pandas
numpy
python-leetcode
flair
sentence_transformers
pyenchant
ipywidgets